% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alnDotPlot.R
\name{generate_dot_matrix}
\alias{generate_dot_matrix}
\title{Generate dot matrix}
\usage{
generate_dot_matrix(input)
}
\arguments{
\item{input}{directory containing all alignments (DNA/AA), list of fasta files,
or a json file containing all alignments of a single raw DNA/AA sequence.}
}
\value{
a single dot matrix.
}
\description{
This function takes all pairwise alignments and creates a dot matrix
representing the alignment's path.
}
