% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alnDotPlot.R
\name{matrix_to_dataframe}
\alias{matrix_to_dataframe}
\title{Convert dot matrix to data frame}
\usage{
matrix_to_dataframe(mat)
}
\arguments{
\item{mat}{dot matrix.}
}
\value{
a data frame.
}
\description{
The dot matrix can be generated by generate_dot_matrix.
}
\details{
Within the output dataframe, the integer values in either columns indicate
where the real match/gap occurs, the non-integer values indicates the locus
of potential insertion space.
}
