% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alnDotPlot.R
\name{plot_dot_matrix}
\alias{plot_dot_matrix}
\title{Plot dot matrix}
\usage{
plot_dot_matrix(input, use_ggplot = FALSE)
}
\arguments{
\item{input}{directory containing all alignments (DNA/AA), list of fasta files,
or a json file containing all alignments of a single raw DNA/AA sequence.}

\item{use_ggplot}{a logical value indicating whether to use ggplot to create
a raster plot or base R. \code{FALSE} by default.}
}
\value{
A raster plot.
}
\description{
This function plots a dot matrix as a raster image.
}
